pyproject.toml
[project]
name = "meditab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "YudaLegend", email = "a779052016@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.73.1",
    "mcp>=1.27.0",
    "pydantic>=2.13.3",
    "pymongo>=4.17.0",
    "python-dotenv>=1.2.2",
]

[project.scripts]
meditab = "meditab:main"

[build-system]
requires = ["uv_build>=0.11.6,<0.12.0"]
build-backend = "uv_build"
